# lib/api-spec/openapi.yaml
openapi: "3.1.0"
info:
  title: Api
  version: "1.0.0"
  description: CQG Plant Inspection API
servers:
  - url: /api
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Enums ──────────────────────────────────────────────────────────────
  schemas:
...
    TemplateType:
      type: string
      enum: [PRE_USE, ROUTINE, SAFETY, ENVIRONMENTAL, CUSTOM]
    InspectionStatus:
      type: string
      enum: [IN_PROGRESS, COMPLETED]
    InspectionResult:
      type: string
      enum: [PASS, FAIL, MONITOR]
    AnswerResult:
      type: string
      enum: [PASS, FAIL, MONITOR, N_A]
    Priority:
      type: string
      enum: [LOW, MEDIUM, HIGH, URGENT]
    WorkOrderStatus:
      type: string
      enum: [OPEN, IN_PROGRESS, WAITING_PARTS, COMPLETED, VERIFIED]
    PPMFrequency:
      type: string
      enum: [DAILY, WEEKLY, MONTHLY, QUARTERLY, ANNUAL]
...
    SiteRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        code:
          type: string
      required: [id, name, code]
...
    AssetRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        plantId:
          type: string
          nullable: true
      required: [id, name]
    UserRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required: [id, name]
    TemplateRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: "#/components/schemas/TemplateType"
      required: [id, name, type]
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]

    # ── Domain models ────────────────────────────────────────────────────
...
    InspectionTemplate:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        type:
          $ref: "#/components/schemas/TemplateType"
        siteId:
          type: string
          format: uuid
          nullable: true
        isActive:
          type: boolean
        sections:
          type: array
          items:
            $ref: "#/components/schemas/TemplateSection"
        _count:
          type: object
          properties:
            inspections:
              type: integer
      required: [id, name, type, isActive]
    TemplateSection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        title:
          type: string
        order:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/TemplateQuestion"
      required: [id, templateId, title, order]
    TemplateQuestion:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sectionId:
          type: string
          format: uuid
        text:
          type: string
        order:
          type: integer
        requiresEvidenceOnFail:
          type: boolean
        allowMonitor:
          type: boolean
        helpText:
          type: string
          nullable: true
      required: [id, sectionId, text, order, requiresEvidenceOnFail, allowMonitor]
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
          nullable: true
        siteId:
          type: string
          format: uuid
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: "#/components/schemas/InspectionStatus"
        overallResult:
          $ref: "#/components/schemas/InspectionResult"
          nullable: true
        locationLat:
          type: number
          nullable: true
        locationLng:
          type: number
          nullable: true
        template:
          $ref: "#/components/schemas/TemplateRef"
          nullable: true
        asset:
          $ref: "#/components/schemas/AssetRef"
          nullable: true
        site:
          $ref: "#/components/schemas/SiteRef"
          nullable: true
        completedBy:
          $ref: "#/components/schemas/UserRef"
          nullable: true
        _count:
          type: object
          properties:
            answers:
              type: integer
            workOrders:
              type: integer
      required: [id, templateId, siteId, startedAt, status]
    InspectionDetail:
      allOf:
        - $ref: "#/components/schemas/Inspection"
        - type: object
          properties:
            template:
              $ref: "#/components/schemas/InspectionTemplate"
            answers:
              type: array
              items:
                $ref: "#/components/schemas/InspectionAnswer"
            workOrders:
              type: array
              items:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  priority:
                    $ref: "#/components/schemas/Priority"
                  status:
                    $ref: "#/components/schemas/WorkOrderStatus"
                  createdAt:
                    type: string
                    format: date-time
    InspectionAnswer:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inspectionId:
          type: string
          format: uuid
        questionId:
          type: string
          format: uuid
        result:
          $ref: "#/components/schemas/AnswerResult"
        notes:
          type: string
          nullable: true
        answeredById:
          type: string
          format: uuid
          nullable: true
        answeredAt:
          type: string
          format: date-time
          nullable: true
        locationLat:
          type: number
          nullable: true
        locationLng:
          type: number
          nullable: true
        question:
          type: object
          nullable: true
          properties:
            id:
              type: string
              format: uuid
            text:
              type: string
            order:
              type: integer
            sectionId:
              type: string
              format: uuid
        answeredBy:
          $ref: "#/components/schemas/UserRef"
          nullable: true
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentRef"
      required: [id, inspectionId, questionId, result]
    InspectionSchedule:
      type: object
      properties:
        id:
          type: string
          format: uuid
        templateId:
          type: string
          format: uuid
        assetId:
          type: string
          format: uuid
          nullable: true
        siteId:
          type: string
          format: uuid
        frequency:
          $ref: "#/components/schemas/PPMFrequency"
        nextDueAt:
          type: string
          format: date-time
        assignedRoleId:
          type: string
          nullable: true
        template:
          $ref: "#/components/schemas/TemplateRef"
          nullable: true
      required: [id, templateId, siteId, frequency, nextDueAt]
...
    AttachmentRef:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        mimeType:
          type: string
        size:
          type: integer
        uploadedAt:
          type: string
          format: date-time
      required: [id, filename, mimeType, size, uploadedAt]
...
paths:
...
  /inspections:
    get:
      operationId: getInspections
      tags: [inspections]
      summary: List inspections
      parameters:
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: assetId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/InspectionStatus"
        - name: templateId
          in: query
          schema:
            type: string
            format: uuid
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                properties:
                  inspections:
                    type: array
                    items:
                      $ref: "#/components/schemas/Inspection"
                  total:
                    type: integer
    post:
      operationId: createInspection
      tags: [inspections]
      summary: Start a new inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId, siteId]
              properties:
                templateId:
                  type: string
                  format: uuid
                assetId:
                  type: string
                  format: uuid
                siteId:
                  type: string
                  format: uuid
                locationLat:
                  type: number
                locationLng:
                  type: number
      responses:
        "201":
          description: Created inspection with full template
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionDetail"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inspections/{id}:
    get:
      operationId: getInspection
      tags: [inspections]
      summary: Get inspection by ID with answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Inspection detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inspections/{id}/answers:
    put:
      operationId: saveInspectionAnswers
      tags: [inspections]
      summary: Save or update inspection answers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [answers]
              properties:
                answers:
                  type: array
                  items:
                    type: object
                    required: [questionId, result]
                    properties:
                      questionId:
                        type: string
                        format: uuid
                      result:
                        $ref: "#/components/schemas/AnswerResult"
                      notes:
                        type: string
                      locationLat:
                        type: number
                      locationLng:
                        type: number
      responses:
        "200":
          description: Saved answers
          content:
            application/json:
              schema:
                type: object
                properties:
                  answers:
                    type: array
                    items:
                      $ref: "#/components/schemas/InspectionAnswer"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Inspection not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inspections/{id}/complete:
    post:
      operationId: completeInspection
      tags: [inspections]
      summary: Complete an inspection
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Completed inspection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inspection"
        "400":
          description: Inspection not in progress
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inspections/schedules:
    get:
      operationId: getInspectionSchedules
      tags: [inspections]
      summary: List inspection schedules
      parameters:
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: assetId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of schedules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/InspectionSchedule"
    post:
      operationId: createInspectionSchedule
      tags: [inspections]
      summary: Create an inspection schedule
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [templateId, siteId, frequency, nextDueAt]
              properties:
                templateId:
                  type: string
                  format: uuid
                assetId:
                  type: string
                  format: uuid
                siteId:
                  type: string
                  format: uuid
                frequency:
                  $ref: "#/components/schemas/PPMFrequency"
                nextDueAt:
                  type: string
                  format: date-time
                assignedRoleId:
                  type: string
      responses:
        "201":
          description: Created schedule
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InspectionSchedule"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /inspections/schedules/{id}:
    delete:
      operationId: deleteInspectionSchedule
      tags: [inspections]
      summary: Delete an inspection schedule
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: Deleted

  # ── Work Orders ──────────────────────────────────────────────────────
...